Put back sending of the snapshot updates in the evaluator
